pyproject.toml
[project]
name = "deep-live-cam"
version = "2.0.3"
description = "Real-time face swap and video deepfake application"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "numpy>=1.23.5,<2",
    "typing-extensions>=4.8.0",
    "opencv-python==4.10.0.84",
    "cv2_enumerate_cameras==1.1.15",
    "onnx==1.18.0",
    "insightface==0.7.3",
    "psutil==5.9.8",
    "tk==0.1.0",
    "customtkinter==5.2.2",
    "pillow==11.1.0",
    "torch; sys_platform == 'darwin'",
    "torch==2.8.0+cu128; sys_platform != 'darwin'",
    "torchvision; sys_platform == 'darwin'",
    "torchvision; sys_platform != 'darwin'",
    "onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "onnxruntime-gpu==1.22.0; sys_platform != 'darwin'",
    "tensorflow; sys_platform != 'darwin'",
    "opennsfw2==0.10.2",
    "protobuf==4.25.1",
    "basicsr @ git+https://github.com/xinntao/BasicSR.git@master",
    "gfpgan @ git+https://github.com/TencentARC/GFPGAN.git@master",
    "pygrabber; sys_platform == 'win32'",
    "coremltools>=9.0 ; sys_platform == 'darwin'",
    "onnx2torch>=1.5.15 ; sys_platform == 'darwin'",
    "mlx>=0.20.0 ; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["modules"]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
name = "pytorch-cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" },
]
